refactor folder structure, add backend instructions to readme
