+++ frontend/lib/screens/eventpg.dart
@@ -4,8 +4,8 @@
 // and two actions.
 import 'dart:ui';
 import 'package:flutter/material.dart';
-import 'nav_drawer.dart';
-import 'register_button.dart';
+import '../widgets/nav_drawer.dart';
+import '../widgets/register_button.dart';
 
 /// This is the main application widget.
 class EventPg extends StatelessWidget {
