+++ frontend/lib/widgets/register_button.dart
